[CFL] Change CFL DC_PRED from int to int16_t
This changes serves to illustrate that DC_PRED does not
need 32 bits. It only needs the specified pixel bit depth.
Which fits inside a signed 16 bit integer.

+++ av1/common/cfl.c
@@ -271,7 +271,7 @@ static INLINE int cfl_idx_to_alpha(int alpha_idx, int joint_sign,
 static INLINE void cfl_build_prediction_lbd(const int16_t *pred_buf_q3,
                                             uint8_t *dst, int dst_stride,
                                             int width, int height, int alpha_q3,
-                                            int dc_pred) {
+                                            int16_t dc_pred) {
   for (int j = 0; j < height; j++) {
     for (int i = 0; i < width; i++) {
       dst[i] =
@@ -286,7 +286,7 @@ static INLINE void cfl_build_prediction_lbd(const int16_t *pred_buf_q3,
 static INLINE void cfl_build_prediction_hbd(const int16_t *pred_buf_q3,
                                             uint16_t *dst, int dst_stride,
                                             int width, int height, int alpha_q3,
-                                            int dc_pred, int bit_depth) {
+                                            int16_t dc_pred, int bit_depth) {
   for (int j = 0; j < height; j++) {
     for (int i = 0; i < width; i++) {
       dst[i] = clip_pixel_highbd(
